Fixed the Machine Workbench.
It now works. You can put the items nedeed for the recipe in any of the
three slots, and it'll work.
Fixed the Machine Workbench GUI, the arrow wasn't showing up at the
right place when processing.
You can also put the items in any of the two slots of the metal mixer,
and the recipe will work too.